% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_datapackage.R
\name{extract_datapackage}
\alias{extract_datapackage}
\title{Extract subset of data package resources}
\usage{
extract_datapackage(path, dir, resource_names, ...)
}
\arguments{
\item{path}{path to source data package datapackage.json file}

\item{dir}{folder to save the new data package}

\item{resource_names}{resources to extract}

\item{...}{root level data package properties that should be replaced or added}
}
\description{
Create a new data package from a subset of the resources of a larger
data package.
}
\details{
The resources included are all the explicit listed in the \code{resource_names}
argument plus the resources referenced in the foreign keys constraints
of the explicit listed resources, even if they themselves are not listed.

The only copied files aside from \code{datapackage.json} from the source data package
are the ones that appear in the \code{resource.path} property of each resource.

External table schema and dialect are dereferenced before the creation of
the target \code{datapackage.json} but the external files are not copied.

\code{dir} need not to exist beforehand and all existing files will be overwritten.
}
\examples{
\dontrun{
extract_datapackage("datapackage.json",
                    dir = "~/my-package/",
                    resource_names = c("ft_convenio", "dm_tempo_anual"),
                    name = "convenio-entrada",
                    title = "Convênios de Entrada",
                    version = "0.1.0",
                    homepage = "https://example.com")
}
}
