% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_reorder.R
\name{list_reorder}
\alias{list_reorder}
\title{Reorder list elements}
\usage{
list_reorder(x, sort_order)
}
\arguments{
\item{x}{a named list.}

\item{sort_order}{a character vector with the elements new order.}
}
\value{
a list with the same elements but in a rearranged order.
}
\description{
Reorder the elements of a named list by the given names
}
\details{
This function reorder the given named list by the given character vector
passed in \code{sort_order}.

The elements that are present in the original list and not present in
\code{sort_order} are moved to the end of the list.
}
\examples{
x <- list(a = 1, b = 2)
list_reorder(x, c("b", "a"))
list_reorder(x, c("b"))
list_reorder(x, c("c"))
}
