% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_resource_db.R
\name{parse_sql}
\alias{parse_sql}
\title{Render SQL query to extract resource from template}
\usage{
parse_sql(datapackage_path, resource_name)
}
\description{

}
\details{

}
\examples{
\dontrun{
parse_sql("datapackage.json", "dm_favorecido")
}
}
