% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/style_tableschema.R
\name{style_tableschema}
\alias{style_tableschema}
\title{Style a yaml table schema}
\usage{
style_tableschema(input, output)
}
\arguments{
\item{input}{yaml file with a table schema.}

\item{output}{output file to store reordered table schema}
}
\description{
Reorder the properties of a \href{https://specs.frictionlessdata.io/table-schema/}{table schema}
serialized as \href{https://yaml.org/}{yaml} overwriting the given file.
}
\details{
The original file is overwritten without user confirmation.
}
\examples{
\dontrun{style_tableschema('schemas/schema.yaml')}
}
